postgres_to_elastic/requirements.txt
certifi==2020.12.5
elasticsearch==7.12.0
psycopg2-binary==2.8.6
urllib3==1.26.4
